+++ version_history.md
@@ -1,5 +1,13 @@
 #Version history#
 
+##1.1 (08/04/12)
+* Removed <tt>_.contains</tt> as alias to <tt>_.include</tt>
+* Added <tt>_.removeRange</tt> (as Array function)
+* Added <tt>_.sameKeys</tt> and <tt>_.contains</tt> (as Collection functions)
+* Added <tt>_.bind</tt> (as Utility function)
+* Updated Moses_Lib_Test.lua
+* Updated documentation
+
 ##1.0 (08/02/12)
 * Added <tt>_.append</tt>, <tt>_.invert</tt>, <tt>_.import</tt>, <tt>_.template</tt>, <tt>_.curry</tt>
 * Updated Moses_Lib_Test.lua